[Refac] 위젯 앞에 RS 이름 추가 및 각종 파일 코드 컨벤션에 맞게 이름 및 파일 내용 수정
