--- Content/Blueprints/Actor/Dungeon/BP_RSWeaponSpawnPadActor.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a60d4124e98da9322e2e8d3dbaae1dfa3b3ec4fc87a4ca9dfab9293ca0d2609a
-size 32161
