+++ Content/Blueprints/Actor/Dungeon/BP_WeaponSpawnPadActor.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b2c9672b1b94096aebafcb48730b9a963d6e5f9e65038597f84a832d9ed628f1
+size 31276
