--- Content/Blueprints/Actor/DungeonNpc/BP_RSDunShopNpcActor.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:486c5e0567290c358c419afbe0f4a38cae07df867ede164b576c39f069be74a8
-size 30635
