+++ Content/Blueprints/Widgets/DunShop/WBP_DunItemWidget.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:478eba1b5f8b6e360f996f79ca61f74bd6c1a1a9869a9e4252cb99c1520aff8d
-size 35434
+oid sha256:cc9e914d0dc121e4b16d3e958b82350c5af273314b66be24ed8d0a7902973478
+size 35539
