+++ Content/Blueprints/Widgets/DunShop/WBP_DunShopWidget.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9a005252da207994116996530588c457129b45e2544e17e6bc5433967e07d0de
-size 37224
+oid sha256:51fa804b13f7a6803633842846699d7e524a4d6379a067eb5aa08d7f903d3574
+size 37398
