+++ Source/RogShop/CheatManager/RSCheatManager.h
@@ -38,13 +38,13 @@ public:
 	UFUNCTION(exec)
 	void SpawnWeaponPad();
 
-	UFUNCTION(exec)
-	void ShowRSDunMainWidget();
-
-public:
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "SpawnMonster")
 	TMap<FString, TSubclassOf<AActor>> MonsterMap;
 
-	UPROPERTY(EditAnywhere, Category = "Spawn")
+private:
+	UPROPERTY(EditDefaultsOnly, Category = "Spawn")
 	TSubclassOf<AActor> WeaponSpawnPadBPClass;
+
+	UPROPERTY(EditDefaultsOnly, Category = "Spawn")
+	TSubclassOf<AActor> DunShopNPCClass;
 };
